Pie chart of labeled EPG data

The function plot_pie allows one to visualize EPG waveforms in a pie chart format.

Usage

plot_pie(data, pietype = c("time", "number"),
waveforms = c("non-probing", "C", "E1", "E2", "G", "pd1", "pd2", "pd"))

Arguments

data

A dataframe containing three columns: time, volts, waveform.

pietype

Select the type of pie chart to display. Default is time - a pie chart showing the percent of time spent in each waveform is displayed. If "number" is selected, then the number of instances for each waveform type is displayed - e.g. if there are 20 distinct waveforms, and 5 are of type C, then C will be 5/20 or 25%.

To avoid double counting pds, the category pd itself (i.e. not pd1 or pd2) only consists of times where pd, not classified into subforms, appears.

waveforms

A list of waveforms to include in the plot. Default is all.

Value

A plotly pie chart of EPG waveforms in a dataset. If none of the specified waveforms are present in the data, then the function will return NULL.

Details

The function plot_pie is designed to help visualize EPG waveform data.

See also