Plot of EPG data colored by waveform

The function plot_wave plots EPG data colored by waveform labels.

Usage

plot_wave(data, facetted = FALSE, aggregate = c("all", "smart", "none"),
interval = 100, smart_int = 10)

Arguments

data

A dataframe containing three columns: time, volts, waveform.

facetted

TRUE/FALSE indicating whether to plot each waveform individually.

aggregate

Optional aggregation parameter, must be one of "all", "smart", or "none". Defaults to all - data will be aggregated on the chosen interval. If "smart" is chosen, certain regions will be aggregated more strongly than the interval period, e.g. non-probing areas. This is recommended to speed up plotting without losing too much resolution in important waveform areas. Lastly, option none results in no aggregation being performed - all data is plotted.

interval

Integer indicating the number of points to aggregate together if aggregate = TRUE.

smart_int

If aggregate = "smart", choose how strongly to further aggregate non-pd sections of the data. Default is 10 for balance between speed and resolution.

Value

A plot of EPG data colored by waveform.

Details

The function plot_wave relies on specific waveform labels which are as follows: a, C, E1, E2, g, pd1, pd2, pd, and NA OR non-probing for non-feeding activity. The labels must be characters and are case sensitive.

See also